Refactor mislocated components
