+++ pluginjar-plugin/src/main/java/kr/junhyung/pluginjar/plugin/ProcessPluginResources.kt
@@ -1,6 +1,6 @@
 package kr.junhyung.pluginjar.plugin
 
-import kr.junhyung.pluginjar.plugin.service.PluginMetaService
+import kr.junhyung.pluginjar.core.service.PluginMetaService
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
 import org.gradle.api.file.DirectoryProperty
