+++ pluginjar-plugin/src/main/java/kr/junhyung/pluginjar/plugin/dsl/BukkitDependencyExtensions.kt
@@ -2,13 +2,7 @@ package kr.junhyung.pluginjar.plugin.dsl
 
 import kr.junhyung.pluginjar.plugin.BukkitDependency
 import kr.junhyung.pluginjar.plugin.BukkitVersion
-import org.gradle.api.JavaVersion
-import org.gradle.api.Project
 import org.gradle.api.artifacts.dsl.DependencyHandler
-import org.gradle.api.plugins.ExtensionAware
-import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.internal.jvm.Jvm
-import org.gradle.kotlin.dsl.getByType
 
 fun DependencyHandler.spigot(version: String): Any {
     return buildBukkitDependency(BukkitDependency.SPIGOT, version)
